+++ src/Exceptions/IdentityConflictException.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Jmluang\SsoConsumer\Exceptions;
+
+class IdentityConflictException extends SsoConsumerException
+{
+    public const ERROR_CODE = 'identity_conflict';
+}